Move constants into a shoes.rb to start off shoes.rb as separate
