+++ examples/button_go_away.rb
@@ -1,4 +1,4 @@
-Shoes.app(debug: true) do
+Shoes.app do
   @b = button "Go away, button"
   @b.click {
     @b.destroy
